switch_interface descriptions
Made lots of changes to switch_interface_descriptions so that
files are saved to a SQLite database.

Have also re-created the script that gets the data from the
SQLLite database and generates an Excel spreadsheet from it.
The speed of generation of the spreadsheet has been significantly
improved as I fixed a pre-existing bug